Return the exact style when you search for '320x_2x'
Change DesiredThumbnail so that is does not understand retina
prefixes aka they are incompatile. By doing this when you do something
like user.artwork_url(:avatar, '320x_2x') you will always get the url
for the 320x_2x style.
Move the logic that scales the width into the current resolution in
DesiredThumbnail, because it contains the base_resolution.
Refactor a bit the DesiredThumbnail.
Add a bit of specs.
Add Artwork::expected_width_for that is a proxy to DesiredThumbnail#width_in_current_resolution